+++ cmp/viewmanager/ViewManagerModel.ts
@@ -129,7 +129,7 @@ export interface ViewManagerConfig {
      * ViewManager implementation to be routable, and you want users to be able to Ctrl+Click or
      * ContextMenu Click to open views from the view menu in new tabs or windows.
      */
-    viewRouteParam?: string;
+    customViewMenuItem?: (view: ViewInfo) => ReactNode;
 }
 
 /**
@@ -172,7 +172,7 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
     readonly instance: string;
     readonly typeDisplayName: string;
     readonly globalDisplayName: string;
-    readonly viewRouteParam: string;
+    readonly customViewMenuItem: (view: ViewInfo) => ReactNode;
     readonly enableAutoSave: boolean;
     readonly enableDefault: boolean;
     readonly enableGlobal: boolean;
@@ -291,7 +291,7 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
         instance = 'default',
         typeDisplayName,
         globalDisplayName = 'global',
-        viewRouteParam,
+        customViewMenuItem,
         manageGlobal = false,
         enableAutoSave = true,
         enableDefault = true,
@@ -308,16 +308,11 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
             "ViewManagerModel requires 'initialViewSpec' if 'enableDefault' is false."
         );
 
-        throwIf(
-            viewRouteParam && isNil(XH.routerState),
-            "Cannot use 'viewRouteParam' if your app does not define a router."
-        );
-
         this.type = type;
         this.instance = instance;
         this.typeDisplayName = lowerCase(typeDisplayName ?? genDisplayName(type));
         this.globalDisplayName = globalDisplayName;
-        this.viewRouteParam = viewRouteParam;
+        this.customViewMenuItem = customViewMenuItem;
         this.manageGlobal = executeIfFunction(manageGlobal) ?? false;
         this.enableDefault = enableDefault;
         this.enableGlobal = enableGlobal;